Fix : #32 `CommandAcceptanceException` 예외 발생 수정

+++ src/main/java/com/toy/sakila/address/application/port/in/AddressCreationCommand.java
@@ -10,7 +10,7 @@ public class AddressCreationCommand {
     String address;
     String address2;
     String district;
-    Long cityId;
+    Short cityId;
     String postalCode;
     String phone;
 }
