+++ src/main/java/com/toy/sakila/address/application/port/in/AddressUpdateCommand.java
@@ -13,7 +13,7 @@ public class AddressUpdateCommand {
     String address;
     String address2;
     String district;
-    Long cityId;
+    Short cityId;
     String postalCode;
     String phone;
 }
