+++ src/main/java/com/toy/sakila/film/adapter/in/in/api/FilmCreationController.java
@@ -46,7 +46,7 @@ public class FilmCreationController {
 
     @Builder
     private record OutputDTO(
-            Long id,
+            Short id,
             String title,
             String description,
             Short releaseYear,
